// src/api/generated/model/decisionStatus.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * PurposePunch.Api
 * OpenAPI spec version: 1.0
 */

export type DecisionStatus =
  (typeof DecisionStatus)[keyof typeof DecisionStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DecisionStatus = {
  Active: "Active",
  Reflected: "Reflected",
  Abandoned: "Abandoned",
} as const;