// src/api/generated/model/updateDecisionCommand.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * PurposePunch.Api
 * OpenAPI spec version: 1.0
 */
import type { Visibility } from "./visibility";
import type { SatisfactionScale } from "./satisfactionScale";

export interface UpdateDecisionCommand {
  id?: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  expectedOutcome?: string | null;
  visibility?: Visibility;
  /** @nullable */
  expectedReflectionDate?: string | null;
  /** @nullable */
  actualOutcome?: string | null;
  /** @nullable */
  lessonsLearned?: string | null;
  /** @nullable */
  privateNotes?: string | null;
  satisfaction?: SatisfactionScale;
}